- Fixed problem with opening a connection that was previously closed by sudden server disconnection (bug #33909)
The problem was that a connection that disconnects suddenly goes through all the steps of removing from the connection pool and sets the state to closed but the internal driver object is not set to null.  This causes a second open to try and use the existing driver object which fails.