- Fixed new parameter handling code so that procedures and functions now work   (bug #34699).
The problem was that to implement output parameters with MySQL we have to use user variables but executing a sproc or function on a connection that is not allowing user variables would throw a missing parameter exception.  We fixed that by adding code that will check to see if the parameter starts with the sproc hash and ignoring it if so.