Bug #36027 Memory leak when using output parameter in stored procedure
The reason for the memory leak was that we Connector/Net created a 
session variable with unique name for each stored procedure call. 
Each session variable takes some memory on the server, hence the leak.

The fix is not using unique names for stored procedure, instead we give 
all output or inout variables a common prefix.

 It would be nice to get rid of the prefix too, however
1 ) there is a complicated logic inside resolve/bind that makes it not 
possible at the moment.
2) stored procedure parameter names could in theory  collide 
with user variables names.