Fix performance regression introduced by using stream timeouts:
- Fix the overhead of timing
1) The Stopwatch class in .NET uses QueryPerformanceCounter for measuring.
While this timer is exact, it incures a considerable overhead of a system call.
To fix that, introduce LowResolutionStopwatch, which measures in milliseconds
and does not involve system calls. We do not need exact values when for IO.

2) Only measure real IO overhead.
Prior to this patch overhead of buffering and compression was measured as well
which resulted in considerable bigger timing overhead. This is now fixed by factoring 
out parts of TimedStream that were responsible for buffering into MySqlStream.


- Fix overhead of setting Stream.ReadTimeout and Stream.WriteTimeout.
Setting these stream properties too often  is  expensive, because they do 
syscalls (setsockopt). 

1)The fix is to track the last set value and do nothing if timeout does not need to 
be changed. 

2)Extra performance tweak is not to change the timeout if new value
only slightly differs from the current property value (we allow for 0.1 sec 
discrepancy)

3)Also, read operations now will only set Stream.ReadTimeout and write operations
only WriteTimeout (prior to the patch, each IO operation would set both timeouts)

- Reduce number of IO calls.
Use Stream.Read to read 4 bytes, instead of 4 times Stream.ReadByte() . 


Wtih all fixes applies, the overhead of timing is close to 0.