Bug #57092 Row-By-Row-Update: only first row updated.
The problem here is that after fix for Bug#54863, AcceptChanges 
were called for the full Table,not for the individual row, so in a 
row-by-row Update() modifications after the first one are lost.

Fixed by calling AcceptChanges for each individual row. 

It is a question whether we still need to the workaround for original
 bug#54863 and do AcceptChanges(), because it looks like the underlying
bug in .NET was fixed  in a recent update, and Bug#54863 is no more 
reproducible  even if we remove  AcceptChanges()  from  
MySqlDataAdapter.Update() entirely.

That is, DbDataAdapter.Update() does AcceptChanges() itself  for every row
just like it is documented. (tried in .NET 2.0, 3.0, 3.5 and 4.0)