Fix for MySql #63997, Oracble bug #13733054. a performance issue due to lack of compliant behavior in SessionProvider.
Fix was already provided by the user and its simple. The problematic part was to reproduce it, because to do so this sequence must happen:
1. Page1 (writetable session) calls SessionProvider.GetItemExclusive
2. Page2 (read only session) Calls SessionProvider.GetItem
3. Page1 Calls SessionProvider.SetAndReleaseItemExclusive
4. After both pages return, a response from calling Page1 again is delayed up to ASP.NET executionTimeout or more.
Which is actually a race condition. To provoke the race condition ManualResetEvents are used (Mutexes proved to be unstable,
most probably due to Mutex having thread affinity (HttpRuntime uses a thread pool, so the thread in global.asax is not necessarily the same
than the one in Page_Load)). 

The logic for thread synchronization is spread between global.asax & Page_Loads, this is because the sequence of events/method calls is
1. Application_BeginRequest
2. SessionProvider.GetItemExclusive (for write sessions) or SessionProvider.GetItem (for read only sessions).
3. Page_Load
4. For pages with writable sessions one of
a) SessionProvider.SetAndReleaseItemExclusive (If session was modified in PageLoad, this scenario reproduces the issue)
b) or SessionProvider.ReleaseItemExclusive (If session was not modified in PageLoad).
5. Application_EndRequest 

After the fix, there is still a delay but it is always consistently below ASP.NET executionTimeout, this is measured in the test, as the time
between the first and last page request endings (so the test passes with the fix and fails without it).

The unit test does testing of the whole ASP.NET stack (except the browser). The Test pasess in both VS2010 & VS2008 (as the rest of web unit tests).

List of changes:
- MySql.Web.Tests.SessionTests (new), the unit test implementation.
- MySql.Web.Tests.global.asax, thread synchronization.
- MySql.Web.Tests.web.config (the web config used in runtime of the mini web app).
- MySql.Web.Tests.web_config_src.txt (the real web config, is changed at runtime with connection data 
  and dealt with for minor .NET 2.x vs .NET 4.x differentes).
- MySql.Web.Tests.InitSessionLocking.aspx (a dummy read only page, just to get the sessionid to reuse from url response).
- MySql.Web.Tests.write.aspx (as Page1 in original scenario (steps 1 & 3). The main session writeable web page, has thread sync logic)
- MySql.Web.Tests.read.aspx (as Page2 in original scenario, the session read only page, has thread sync logic).
- MySql.Web.Tests.write2.aspx (as Page1 in original scenario step 4, Another session writetable web page, 
  does not have thread synchronization logic (that's why a "double" is needed)).
- MySql.Web.SessionProvider (the fix itself).