Mysql Bug 70602: .Net connector does not add auto_increment to bigint columns
Description:
In EF5, when using the HasDatabaseGeneratedOption(DatabaseGeneratedOption.Identity) method, the column is not created with the AUTO_INCREMENT directive if the column is an Int64. It works perfectly for an Int32 column. When I looked at the source, I saw that only ints are supported, not bigints or possibly tinyint.

Fix:
- MySqlMigrationSqlGenerator: handle the data types tinyint, smallint, mediumint, int, bigint as possible primary key data type. Support that a column that is auto incremental could be the primary key.
- App.config: add connection string for new test.
- CodeFirstTests: new test added.
- Vehicle: class used by the test created.
- CHANGES: comment with the bug fixed added.