Jira MYSQLFORVS-444: Object Reference error with MySQL MVC Item Template.
Fix:
Item template sections commented temporally to not ship the item templates for VS 2010 and 2012 versions.
This due an exception thrown when creating the item template, the bug will be fixed in a future release so the team member assigned will just uncomment the sections to start working in the fix.